[REFACTOR] enum 위치 변경에 따른 코드 리팩터링
