+++ CarrotServer/src/main/java/CarrotServer/domain/Lives.java
@@ -1,5 +1,6 @@
 package CarrotServer.domain;
 
+import CarrotServer.common.enums.LifeCategory;
 import jakarta.persistence.*;
 import lombok.Getter;
